Clean assembly organization.
