+++ include/gendil/FiniteElementMethod/MatrixAssembly/assembly.hpp
@@ -13,8 +13,9 @@
 // Assembly of block sparse matrices in BSR format.
 #include "gendil/FiniteElementMethod/MatrixAssembly/BSR/bsrassembly.hpp"
 
+// Scatter/gather BSR assembly wrapper for finite-element vectors.
+#include "gendil/FiniteElementMethod/MatrixAssembly/SGBSR/sgbsrassembly.hpp"
+
 // Raw COO triplet sparse assembly (unsorted and duplicated entries).
 #include "gendil/FiniteElementMethod/MatrixAssembly/COO/cooassembly.hpp"
 
-// Scatter/gather wrapper for applying BSR matrices to finite-element vectors.
-#include "gendil/FiniteElementMethod/MatrixAssembly/SGBSR/sgbsrgatherscatter.hpp"
